get rid of favorites 2
