--- shared/actions/json/favorite.json
@@ -1,37 +1,0 @@
-{
-  "prelude": [
-    "import * as Types from '../constants/types/favorite'",
-    "import * as Folders from '../constants/types/folders'"
-  ],
-  "actions": {
-    "favoriteAdd": {
-      "path": "string"
-    },
-    "favoriteAdded": {
-      "canError": {
-        "errorText": "string"
-      }
-    },
-    "favoriteIgnore": {
-      "path": "string"
-    },
-    "favoriteIgnored": {
-      "canError": {
-        "errorText": "string"
-      }
-    },
-    "favoriteList": {},
-    "favoriteListed": {
-      "folders": "Types.FolderState"
-    },
-    "favoriteSwitchTab": {
-      "showingPrivate": "boolean"
-    },
-    "favoriteToggleIgnored": {
-      "isPrivate": "boolean"
-    },
-    "markTLFCreated": {
-      "folder": "Folders.Folder"
-    }
-  }
-}
