+++ shared/constants/reducer.js
@@ -4,7 +4,6 @@ import type {State as Config} from '../constants/types/config'
 import type {State as Dev} from '../constants/types/dev'
 import type {State as Devices} from '../constants/types/devices'
 import type {State as Entity} from '../constants/types/entities'
-import type {State as Favorite} from '../constants/types/favorite'
 import type {State as Git} from '../constants/types/git'
 import type {State as Gregor} from '../constants/types/gregor'
 import type {State as FS} from '../constants/types/fs'
@@ -31,7 +30,6 @@ export type TypedState = $ReadOnly<{|
   dev: Dev,
   devices: Devices,
   entities: Entity,
-  favorite: Favorite,
   fs: FS,
   git: Git,
   gregor: Gregor,
