+++ shared/store/configure-sagas.js
@@ -3,7 +3,6 @@ import chat2Saga from '../actions/chat2'
 import configSaga from '../actions/config'
 import createSagaMiddleware from 'redux-saga'
 import deviceSaga from '../actions/devices'
-import favoriteSaga from '../actions/favorite'
 import fsSaga from '../actions/fs'
 import gitSaga from '../actions/git'
 import gregorSaga from '../actions/gregor'
@@ -32,7 +31,6 @@ function* mainSaga(): Saga.SagaGenerator<any, any> {
   yield Saga.fork(chat2Saga)
   yield Saga.fork(configSaga)
   yield Saga.fork(deviceSaga)
-  yield Saga.fork(favoriteSaga)
   yield Saga.fork(fsSaga)
   yield Saga.fork(gregorSaga)
   yield Saga.fork(kbfsSaga)
